requirements.txt
requests
pillow
beautifulsoup4
easyocr
pytesseract
